Optimize carrying over notes after issuing a note cut in conversions
The note parameter is not overwritten when a note cut is issued on the SPC side,
especially since a note cut is issued with a zero flag (along with a zero
instrument ID). Therefore, the conversion can take advantage of this detail by
not factoring in a note cut when recording the last note used in a track.